[BUS-806] refactor nameing to match conventions
