+++ pages/account/address.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import Layout from '~/components/Layout/Account/BaseLayout.vue';
-import AddressContent from '~/components/Layout/Account/AdressContent.vue';
+import LayoutAccount from '~/components/Layout/Account/LayoutAccount.vue';
+import AddressContent from '~/components/Layout/Account/LayoutAccountAdressContent.vue';
 </script>
 <template>
-    <layout>
+    <layout-account>
         <address-content/>
-    </layout>
+    </layout-account>
 </template>
 
 <style scoped>
