--- pages/account/overview.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-<div>Hello Overview</div>
-</template>
-
-<style scoped>
-
-</style>
